Fix bug in connection search
Previously, the results of the connection search query were not sorted
by any specific criterion. This lead to unpredictable bugs when
querying the route between two train stations with at least two
different train connections required.

This commit sorts the query results by the search path in ascending
order to guarantee that the search tree is always constructed in the
same manner.